requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.31.0
openai==1.12.0
chromadb==0.4.22
sqlalchemy==2.0.25
python-multipart==0.0.9
python-dotenv==1.0.1
reportlab==4.0.9
pandas==2.2.0
pydantic==2.6.1
requests==2.31.0
aiofiles==23.2.1
tenacity==8.2.3
passlib==1.7.4
bcrypt==4.1.2
jose==1.0.0
onnxruntime==1.23.2
